requirements.txt
streamlit==1.31.0
pandas==2.2.0
plotly==5.18.0
openpyxl==3.1.2
python-dateutil==2.8.2
google-generativeai
